refactor(#114): like api 분리
