+++ src/main/java/deepdivers/community/domain/comment/controller/docs/CommentApiControllerDocs.java
@@ -91,38 +91,4 @@ public interface CommentApiControllerDocs {
     )
     ResponseEntity<NoContent> removeCommentOnPost(Member member, RemoveCommentRequest request);
 
-    @Operation(summary = "댓글 및 답글 좋아요", description = "댓글 및 답글을 좋아요 하는 기능")
-    @ApiResponse(
-        responseCode = "1406",
-        description = """
-                    1. 댓글 좋아요에 성공했습니다.
-                    """
-    )
-    @ApiResponse(
-        responseCode = "2500\n9000~9005",
-        description = """
-            1. 유효하지 않은 접근입니다.
-            2. 토큰 관련 예외입니다.
-            """,
-        content = @Content(schema = @Schema(implementation = ExceptionResponse.class))
-    )
-    ResponseEntity<NoContent> likeComment(Member member, LikeRequest request);
-
-    @Operation(summary = "댓글 및 답글 좋아요 취소", description = "댓글 및 답글을 좋아요를 취소 하는 기능")
-    @ApiResponse(
-        responseCode = "1407",
-        description = """
-                    1. 댓글 좋아요 취소에 성공했습니다.
-                    """
-    )
-    @ApiResponse(
-        responseCode = "2500\n9000~9005",
-        description = """
-            1. 유효하지 않은 접근입니다.
-            2. 토큰 관련 예외입니다.
-            """,
-        content = @Content(schema = @Schema(implementation = ExceptionResponse.class))
-    )
-    ResponseEntity<NoContent> unlikeComment(Member member, LikeRequest request);
-
 }
