+++ src/main/java/deepdivers/community/domain/post/controller/docs/PostApiControllerDocs.java
@@ -41,35 +41,5 @@ public interface PostApiControllerDocs {
 	)
 	ResponseEntity<NoContent> deletePost(Member member, Long postId);
 
-	@Operation(summary = "게시글 좋아요", description = "게시글 좋아요를 하는 기능")
-	@ApiResponse(
-			responseCode = "1205",
-			description = " 게시글 좋아요에 성공했습니다."
-	)
-	@ApiResponse(
-			responseCode = "2500\n9000~9005",
-			description = """
-            1. 유효하지 않은 접근입니다.
-            2. 토큰 관련 예외입니다.
-            """,
-			content = @Content(schema = @Schema(implementation = ExceptionResponse.class))
-	)
-	ResponseEntity<NoContent> likePost(Member member, LikeRequest request);
-
-	@Operation(summary = "게시글 좋아요 취소", description = "게시글 좋아요를 취소 하는 기능")
-	@ApiResponse(
-			responseCode = "1206",
-			description = "게시글 좋아요 취소에 성공했습니다."
-	)
-	@ApiResponse(
-			responseCode = "2500\n9000~9005",
-			description = """
-            1. 유효하지 않은 접근입니다.
-            2. 토큰 관련 예외입니다.
-            """,
-			content = @Content(schema = @Schema(implementation = ExceptionResponse.class))
-	)
-	ResponseEntity<NoContent> unlikePost(Member member, LikeRequest request);
-
 }
